refactor(inference): extract provider clients into parish-providers crate
Split parish-inference along the transport/scheduling boundary. The provider
HTTP clients (openai_client/, anthropic_client/), the offline simulator + mock
backends, the AnyClient dispatch enum (any_client.rs), the outbound rate
limiter (rate_limit.rs, governor), and the shared client_base.rs/utf8_stream.rs
move into a new backend-agnostic workspace crate, parish-providers.
parish-inference keeps queue/worker/timeout/validate/file_log/setup and now
depends on parish-providers, re-exporting every moved symbol at its former
parish_inference::* path so no downstream consumer (parish-core, -npc, -input,
-tauri, -engine) changes a single import.

- New crate parish/crates/parish-providers/ (Cargo.toml + README + lib.rs);
  workspace members + [workspace.dependencies] updated.
- governor moves to parish-providers; reqwest's streaming clients move with it
  (parish-inference keeps a json-only reqwest for validate/setup/hf_downloader
  HTTP probes).
- StreamStats (any_client) and build_client_or_fallback (openai_client) widened
  pub(crate) -> pub so the staying worker/setup code can use them cross-crate.
- architecture_fitness BACKEND_AGNOSTIC list + reqwest-containment docs updated
  to name parish-providers; no #[allow] added, no dependency cycle.
- Docs: architecture.md, codebase-map.md, README, AGENTS.md crate counts
  (17 -> 18) + leaf-crate lists; parish-inference scope doc; THIRD_PARTY_NOTICES
  regenerated.

All existing tests pass unmodified (parish-providers 141, parish-inference 145,
parish-core rate-limit + arch-fitness). Simulator dialogue fixture output is
structurally identical before/after.
