+++ LEARNINGS.md
@@ -72,3 +72,4 @@ bottom; don't lengthen items past 2-3 lines.
 - **`parish-server`'s file logging is effectively dead — `tracing::debug!`/`info!` from the binary never reach disk.** `setup_tracing_and_otel` in `parish-server/src/main.rs` builds a `tracing_appender::non_blocking` writer but binds its worker guard to `let (_non_blocking, _guard)` — the `_guard` is dropped at the end of that function, so the background writer thread shuts down and `logs/parish-server.log.<date>` stays 0 bytes. Consequence for live proofs: you cannot grep the server log to observe an async `/api/submit-input` code path (e.g. a turn dropped by the #1379 serialization guard). Workarounds: drive the **synchronous** `/api/command` path via the `parish` client (it returns the full turn body, incl. a `409 {"reason":"command_in_progress"}` rejection), or assert via unit/integration tests — there is no host-visible transcript endpoint (`engine-state` omits the conversation transcript; `debug-snapshot.conversations` is live-chain state that resets per turn, exchange_count stays 0). Likely a real bug worth fixing (move `_guard` to a `'static`/returned binding).
 - **A worktree's branch may be locked to a _different_ worktree.** If `git checkout <branch>` fails with `'<branch>' is already used by worktree at <path>`, the orchestrator pre-created that branch in another agent worktree; that sibling may have an _untracked_ fixture but **no commits / nothing pushed**. Don't fight the lock — create a distinct local branch (e.g. `<branch>-impl`) in your own worktree and open the PR from it. (Hit during #1379.)
 - **NPC `dialogue` is served by the BASE provider/model slot, NOT per-category client resolution.** Only `intent` / `reaction` / `simulation` route through `GameConfig::resolve_category_client` (which reads `category_provider/model/base_url`). The dialogue path in `parish-core/game_loop/npc_turn.rs` uses `config.model_name` + the single base worker built by `rebuild_inference_worker` (which only takes base provider/url/key). Consequence: setting `/provider.dialogue` / `/model.dialogue` updates the debug-snapshot `categories[dialogue]` entry but the actual dialogue inference STILL uses the base model — the snapshot lies. To BYOK the dialogue model you must change the BASE slot: `/provider`, `/url`, `/model`, `/key` (no `.dialogue` suffix). The harness `apply_byok` (#1365) special-cases `category=="dialogue"` → base commands, all others → dotted. Verify a rebind actually took effect by reading `inference.call_log[].model` after a real turn, never just the `categories` snapshot. Also: base `/url` (`SetBaseUrl`) and `/url.<cat>` (`SetCategoryBaseUrl`) commands were added in #1365 — and base `SetModel` / `SetCategoryModel` now emit `RebuildInference` (they previously did NOT, so a model-only change silently didn't rebind the worker).
+- **`parish-inference` is split into transport (`parish-providers`) + scheduling (`parish-inference`).** The provider HTTP clients (`openai_client/`, `anthropic_client/`), `simulator.rs`, `mock_client.rs`, `any_client.rs` (`AnyClient`, `build_client`, `TOKEN_CHANNEL_CAPACITY`, `StreamStats`), `rate_limit.rs` (`InferenceRateLimiter` + `governor`), `client_base.rs`, and `utf8_stream.rs` live in `parish-providers`; queue/worker/timeout/validate/file_log/setup/hf_downloader/Ollama-`client.rs` stay in `parish-inference`. `parish-inference` re-exports EVERY moved symbol (incl. the whole `any_client`/`openai_client`/`simulator`/… modules) at the unchanged `parish_inference::*` path, so downstream (`parish-core`'s `pub use parish_inference as inference`, npc, input, tauri, engine) needs zero import changes. Gotchas when touching this: (1) `parish-inference` STILL needs a direct `reqwest` dep (json feature) — `validate.rs`/`hf_downloader.rs`/`setup/` make their own HTTP calls, only the `stream` feature + streaming clients are exclusive to `parish-providers`; (2) `build_client_or_fallback` (openai_client) and `StreamStats` (any_client) are `pub` (not `pub(crate)`) precisely because `parish-inference::setup`/`worker` construct them cross-crate — don't re-narrow them; (3) `parish-providers` must never depend on `parish-inference` (cycle) and is in the `BACKEND_AGNOSTIC` arch-fitness list. The `--script` headless harness (`GameTestHarness::build_with_mod`) does NOT enable the simulator fallback and reports provider as `ollama`; drive deterministic NPC dialogue proofs with `/stub <name>: <text>` (→ `npc_response` with real text), not by expecting a `provider:simulator` line.
