+++ parish/Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "crates/parish-geo-tool",
     "crates/parish-types",
     "crates/parish-config",
+    "crates/parish-providers",
     "crates/parish-inference",
     "crates/parish-world",
     "crates/parish-palette",
@@ -39,6 +40,7 @@ default-members = ["crates/parish-engine", "crates/parish-mcp"]
 # Internal crates
 parish-types = { path = "crates/parish-types" }
 parish-config = { path = "crates/parish-config" }
+parish-providers = { path = "crates/parish-providers" }
 parish-inference = { path = "crates/parish-inference" }
 parish-world = { path = "crates/parish-world" }
 parish-palette = { path = "crates/parish-palette" }
