+++ parish/THIRD_PARTY_NOTICES.ui.md
@@ -1,68 +1,1 @@
-- [@esbuild/darwin-arm64@0.27.4](https://github.com/evanw/esbuild) - MIT
-- [@jridgewell/gen-mapping@0.3.13](https://github.com/jridgewell/sourcemaps) - MIT
-- [@jridgewell/remapping@2.3.5](https://github.com/jridgewell/sourcemaps) - MIT
-- [@jridgewell/resolve-uri@3.1.2](https://github.com/jridgewell/resolve-uri) - MIT
-- [@jridgewell/sourcemap-codec@1.5.5](https://github.com/jridgewell/sourcemaps) - MIT
-- [@jridgewell/trace-mapping@0.3.31](https://github.com/jridgewell/sourcemaps) - MIT
-- [@mapbox/jsonlint-lines-primitives@2.0.2](https://github.com/mapbox/jsonlint) - MIT
-- [@mapbox/point-geometry@1.1.0](https://github.com/mapbox/point-geometry) - ISC
-- [@mapbox/tiny-sdf@2.1.0](https://github.com/mapbox/tiny-sdf) - BSD-2-Clause
-- [@mapbox/unitbezier@0.0.1](https://github.com/mapbox/unitbezier) - BSD-2-Clause
-- [@mapbox/vector-tile@2.0.4](https://github.com/mapbox/vector-tile-js) - BSD-3-Clause
-- [@mapbox/whoots-js@3.1.0](https://github.com/mapbox/whoots-js) - ISC
-- [@maplibre/geojson-vt@5.0.4](https://github.com/maplibre/geojson-vt) - ISC
-- [@maplibre/geojson-vt@6.0.4](https://github.com/maplibre/geojson-vt) - ISC
-- [@maplibre/maplibre-gl-style-spec@24.8.1](https://github.com/maplibre/maplibre-style-spec) - ISC
-- [@maplibre/mlt@1.1.8](https://github.com/maplibre/maplibre-tile-spec) - (MIT OR Apache-2.0)
-- [@maplibre/vt-pbf@4.3.0](https://github.com/maplibre/vt-pbf) - MIT
-- [@rollup/rollup-darwin-arm64@4.60.0](https://github.com/rollup/rollup) - MIT
-- [@sveltejs/acorn-typescript@1.0.9](https://github.com/sveltejs/acorn-typescript) - MIT
-- [@tauri-apps/api@2.11.0](https://github.com/tauri-apps/tauri) - Apache-2.0 OR MIT
-- [@types/estree@1.0.8](https://github.com/DefinitelyTyped/DefinitelyTyped) - MIT
-- [@types/geojson@7946.0.16](https://github.com/DefinitelyTyped/DefinitelyTyped) - MIT
-- [@types/node@25.6.0](https://github.com/DefinitelyTyped/DefinitelyTyped) - MIT
-- [@types/supercluster@7.1.3](https://github.com/DefinitelyTyped/DefinitelyTyped) - MIT
-- [@types/trusted-types@2.0.7](https://github.com/DefinitelyTyped/DefinitelyTyped) - MIT
-- [@typescript-eslint/types@8.57.2](https://github.com/typescript-eslint/typescript-eslint) - MIT
-- [acorn@8.16.0](https://github.com/acornjs/acorn) - MIT
-- [aria-query@5.3.1](https://github.com/A11yance/aria-query) - Apache-2.0
-- [axobject-query@4.1.0](https://github.com/A11yance/axobject-query) - Apache-2.0
-- [clsx@2.1.1](https://github.com/lukeed/clsx) - MIT
-- [devalue@5.8.1](https://github.com/sveltejs/devalue) - MIT
-- [earcut@3.0.2](https://github.com/mapbox/earcut) - ISC
-- [esbuild@0.27.4](https://github.com/evanw/esbuild) - MIT
-- [esm-env@1.2.2](https://github.com/benmccann/esm-env) - MIT
-- [esrap@2.2.4](https://github.com/sveltejs/esrap) - MIT
-- [estree-walker@3.0.3](https://github.com/Rich-Harris/estree-walker) - MIT
-- [fdir@6.5.0](https://github.com/thecodrr/fdir) - MIT
-- [fsevents@2.3.3](https://github.com/fsevents/fsevents) - MIT
-- [gl-matrix@3.4.4](https://github.com/toji/gl-matrix) - MIT
-- [html-to-image@1.11.13](https://github.com/bubkoo/html-to-image) - MIT
-- [is-reference@3.0.3](https://github.com/Rich-Harris/is-reference) - MIT
-- [json-stringify-pretty-compact@4.0.0](https://github.com/lydell/json-stringify-pretty-compact) - MIT
-- [kdbush@4.0.2](https://github.com/mourner/kdbush) - ISC
-- [locate-character@3.0.0](git+https://gitlab.com/Rich-Harris/locate-character) - MIT
-- [magic-string@0.30.21](https://github.com/Rich-Harris/magic-string) - MIT
-- [maplibre-gl@5.22.0](https://github.com/maplibre/maplibre-gl-js) - BSD-3-Clause
-- [minimist@1.2.8](https://github.com/minimistjs/minimist) - MIT
-- [murmurhash-js@1.0.0](https://github.com/mikolalysenko/murmurhash-js) - MIT
-- [nanoid@3.3.11](https://github.com/ai/nanoid) - MIT
-- [pbf@4.0.1](https://github.com/mapbox/pbf) - BSD-3-Clause
-- [phosphor-svelte@3.1.0](https://github.com/haruaki07/phosphor-svelte) - MIT
-- [picocolors@1.1.1](https://github.com/alexeyraspopov/picocolors) - ISC
-- [picomatch@4.0.4](https://github.com/micromatch/picomatch) - MIT
-- [postcss@8.5.10](https://github.com/postcss/postcss) - MIT
-- [potpack@2.1.0](https://github.com/mapbox/potpack) - ISC
-- [protocol-buffers-schema@3.6.1](https://github.com/mafintosh/protocol-buffers-schema) - MIT
-- [quickselect@3.0.0](https://github.com/mourner/quickselect) - ISC
-- [resolve-protobuf-schema@2.1.0](https://github.com/mafintosh/resolve-protobuf-schema) - MIT
-- [rollup@4.60.0](https://github.com/rollup/rollup) - MIT
-- [rw@1.3.3](https://github.com/mbostock/rw) - BSD-3-Clause
-- [source-map-js@1.2.1](https://github.com/7rulnik/source-map-js) - BSD-3-Clause
-- [supercluster@8.0.1](https://github.com/mapbox/supercluster) - ISC
-- [svelte@5.55.7](https://github.com/sveltejs/svelte) - MIT
-- [tinyglobby@0.2.15](https://github.com/SuperchupuDev/tinyglobby) - MIT
-- [tinyqueue@3.0.0](https://github.com/mourner/tinyqueue) - ISC
-- [undici-types@7.19.2](https://github.com/nodejs/undici) - MIT
-- [vite@7.3.2](https://github.com/vitejs/vite) - MIT
-- [zimmerframe@1.1.4](https://github.com/sveltejs/zimmerframe) - MIT
+
