+++ parish/crates/parish-providers/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "parish-providers"
+version = "0.1.0"
+edition = "2024"
+license = "GPL-3.0-only"
+publish = false
+description = "LLM provider transport clients (OpenAI/Anthropic/simulator/mock) and outbound rate limiting for Parish"
+
+[dependencies]
+parish-types = { workspace = true }
+parish-config = { workspace = true }
+reqwest = { workspace = true, features = ["json", "stream"] }
+tokio = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+tracing = { workspace = true }
+chrono = { workspace = true }
+governor = { workspace = true }
+regex = { workspace = true }
+
+[dev-dependencies]
+wiremock = { workspace = true }
