+++ parish/crates/parish-providers/src/openai_client/wire.rs
